Reverse the dependency between pyk.kastManip and pyk.cterm
