+++ pyk/src/pyk/kcfg.py
@@ -8,17 +8,9 @@ from typing import Any, Container, Dict, Iterable, List, Mapping, Optional, Sequ
 
 from graphviz import Digraph
 
-from .cterm import CTerm
+from .cterm import CTerm, build_claim, build_rule
 from .kast import KClaim, KInner, KRule, Subst
-from .kastManip import (
-    build_claim,
-    build_rule,
-    ml_pred_to_bool,
-    mlAnd,
-    remove_generated_cells,
-    remove_source_attributes,
-    simplify_bool,
-)
+from .kastManip import ml_pred_to_bool, mlAnd, remove_generated_cells, remove_source_attributes, simplify_bool
 from .ktool import KPrint
 from .utils import add_indent, compare_short_hashes, shorten_hash
