+++ pyk/src/pyk/ktool/kit.py
@@ -6,9 +6,9 @@ from pathlib import Path
 from typing import Any, Dict, Final, Iterable, List, Optional
 
 from ..cfg_manager import CFGManager
-from ..cterm import CTerm
+from ..cterm import CTerm, build_claim
 from ..kast import KApply, KDefinition, KFlatModuleList, KInner, KSequence, KToken, Subst
-from ..kastManip import build_claim, collectFreeVars, getCell, is_top, minimize_term
+from ..kastManip import collectFreeVars, getCell, is_top, minimize_term
 from ..kcfg import KCFG
 from ..ktool import KProve
 from ..prelude import Bool, mlAnd, mlEqualsTrue, mlOr, mlTop
