+++ pyk/src/tests/test_kastManip.py
@@ -1,11 +1,8 @@
 from unittest import TestCase
 
-from pyk.cterm import CTerm
-from pyk.kast import DOTS, KApply, KAtt, KClaim, KInner, KLabel, KRewrite, KSequence, KSort, KVariable
+from pyk.kast import DOTS, KApply, KLabel, KRewrite, KSequence, KSort, KVariable
 from pyk.kastManip import (
     bool_to_ml_pred,
-    build_claim,
-    build_rule,
     collapseDots,
     minimize_term,
     ml_pred_to_bool,
@@ -15,26 +12,15 @@ from pyk.kastManip import (
     splitConfigFrom,
     substitute,
 )
-from pyk.prelude import Bool, Sorts, intToken, mlAnd, mlEqualsTrue, mlTop
+from pyk.prelude import Bool, Sorts, intToken, mlEqualsTrue, mlTop
 
-from .utils import a, b, c, f, k
+from .utils import a, b, c, f, k, x
 
-x = KVariable('X')
-mem = KLabel('<mem>')
-
-T = KLabel('<T>')
 K_CELL = KApply('<k>', [KSequence([KVariable('S1'), KVariable('_DotVar0')])])
 T_CELL = KApply('<T>', [K_CELL, KApply('<state>', [KVariable('MAP')])])
 GENERATED_COUNTER_CELL = KApply('<generatedCounter>', [KVariable('X')])
 GENERATED_TOP_CELL_1 = KApply('<generatedTop>', [T_CELL, KVariable('_GENERATED_COUNTER_PLACEHOLDER')])
 GENERATED_TOP_CELL_2 = KApply('<generatedTop>', [T_CELL, GENERATED_COUNTER_CELL])
-
-v1 = KVariable('V1')
-v2 = KVariable('V2')
-unds_v1 = KVariable('_V1')
-unds_v2 = KVariable('_V2')
-ques_v2 = KVariable('?V2')
-ques_unds_v2 = KVariable('?_V2')
 
 
 class PushDownRewritesTest(TestCase):
@@ -46,39 +32,6 @@ class PushDownRewritesTest(TestCase):
             with self.subTest(i=i):
                 # When
                 actual = push_down_rewrites(before)
-
-                # Then
-                self.assertEqual(actual, expected)
-
-
-class BuildRuleTest(TestCase):
-    def test_build_rule(self):
-        # Given
-        test_data = [
-            (
-                T(k(KVariable('K_CELL')), mem(KVariable('MEM_CELL'))),
-                T(
-                    k(KVariable('K_CELL')),
-                    mem(KApply('_[_<-_]', [KVariable('MEM_CELL'), KVariable('KEY'), KVariable('VALUE')])),
-                ),
-                ['K_CELL'],
-                T(
-                    k(KVariable('_K_CELL')),
-                    mem(
-                        KRewrite(
-                            KVariable('MEM_CELL'),
-                            KApply('_[_<-_]', [KVariable('MEM_CELL'), KVariable('?_KEY'), KVariable('?_VALUE')]),
-                        )
-                    ),
-                ),
-            )
-        ]
-
-        for i, (lhs, rhs, keep_vars, expected) in enumerate(test_data):
-            with self.subTest(i=i):
-                # When
-                rule, _ = build_rule(f'test-{i}', CTerm(lhs), CTerm(rhs), keep_vars=keep_vars)
-                actual = rule.body
 
                 # Then
                 self.assertEqual(actual, expected)
@@ -231,39 +184,6 @@ class SimplifyBoolTest(TestCase):
 
                 # Then
                 self.assertEqual(bool_out_actual, bool_out)
-
-
-class BuildClaimtest(TestCase):
-    def test_build_claim(self):
-        # (<k> V1 </k> #And { true #Equals 0 <=Int V2}) => <k> V2 </k>      expected: <k> _V1 => V2 </k> requires 0 <=Int V2
-        # <k> V1 </k> => <k> V2 </k>                                        expected: <k> _V1 => ?_V2 </k>
-        # <k> V1 </k> => <k> V2 </k> #And { true #Equals 0 <=Int V2 }       expected: <k> _V1 => ?V2 </k> ensures 0 <=Int ?V2
-
-        def constraint(v: KVariable) -> KInner:
-            return KApply('_<=Int_', [intToken(0), v])
-
-        test_data = (
-            (
-                'req-rhs',
-                mlAnd([k(v1), mlEqualsTrue(constraint(v2))]),
-                k(v2),
-                KClaim(k(KRewrite(unds_v1, v2)), requires=constraint(v2), att=KAtt({'label': 'claim'})),
-            ),
-            ('free-rhs', k(v1), k(v2), KClaim(k(KRewrite(unds_v1, ques_unds_v2)), att=KAtt({'label': 'claim'}))),
-            (
-                'bound-rhs',
-                k(v1),
-                mlAnd([k(v2), mlEqualsTrue(constraint(v2))]),
-                KClaim(k(KRewrite(unds_v1, ques_v2)), ensures=constraint(ques_v2), att=KAtt({'label': 'claim'})),
-            ),
-        )
-
-        for name, init, target, claim in test_data:
-            with self.subTest(name):
-                init_cterm = CTerm(init)
-                target_cterm = CTerm(target)
-                kclaim, _ = build_claim('claim', init_cterm, target_cterm)
-                self.assertEqual(kclaim, claim)
 
 
 class SplitConfigTest(TestCase):
